+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Added:
 * Added an option to the config.yml to watch the content folder for changes and keep the car search index updated.
 * Added an option to the config.yml to prevent server manager from opening a browser on Windows on start.
 * Accounts are now part of the shared JSON store. This means if you were running split JSON stores using the 'shared_data_path' config.yml variable, you will need to re-set up server manager accounts (i.e. by copying them into the shared JSON store)
+* You can now import existing Custom Races and Race Weekends to Championships.
 
 Fixes: